requirements.txt
Django==1.8.4
react==1.0.0
djangorestframework==3.2.2
pyquery==1.2.9